Overhaul of the graph library.
GRAPH has been split into GRAPH and GRAPH-PATH. The latter has a new Path
sort which preserves the order of the edges. It also fixes a bug in the old
path-finding algorithm, which did not respect the directionality of edges
in the graph (it treated all edges as undirected).

META-GRAPH has been moved to its own file until I can overhaul it.